+++ src/lib/foundations/graphics/parametric_curve_point.cpp
@@ -44,7 +44,7 @@ void parametric_curve_point::init(double t, int f_is_valid,
 		coords.push_back(x[i]);
 	}
 	if (f_v) {
-		for (i = 0; i < coords.size(); i++) {
+		for (i = 0; i < (int) coords.size(); i++) {
 			cout << coords[i];
 			if (i < coords.size() - 1) {
 				cout << ",";
